refactor: general icons
