--- src/assets/icons/ArrowTriangleIcon.tsx
@@ -1,19 +1,0 @@
-export default function ArrowTriangleIcon(props: any) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="12"
-      height="13"
-      fill="none"
-      viewBox="0 0 12 13"
-      {...props}
-    >
-      <path
-        className="fill-current"
-        fillRule="evenodd"
-        d="M2.654 4.731A.375.375 0 0 1 3 4.5h6a.375.375 0 0 1 .265.64l-3 3a.375.375 0 0 1-.53 0l-3-3a.375.375 0 0 1-.081-.409Z"
-        clipRule="evenodd"
-      />
-    </svg>
-  );
-}
