--- src/assets/icons/BoosterIcon.tsx
@@ -1,23 +1,0 @@
-export default function BoosterIcon(props: any) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      fill="none"
-      viewBox="0 0 24 24"
-      {...props}
-    >
-      <path
-        className="fill-current"
-        fillRule="evenodd"
-        d="M12 20.667a8.667 8.667 0 1 0 0-17.334 8.667 8.667 0 0 0 0 17.334Zm0 2c5.89 0 10.666-4.776 10.666-10.667S17.891 1.333 12 1.333 1.333 6.11 1.333 12 6.109 22.667 12 22.667Z"
-        clipRule="evenodd"
-      />
-      <path
-        className="fill-current"
-        d="M9.867 17.333 10.4 13.6H7.733l4.8-6.933H13.6l-.533 4.266h3.2l-5.334 6.4H9.867Z"
-      />
-    </svg>
-  );
-}
