--- src/assets/icons/CheckIcon.tsx
@@ -1,19 +1,0 @@
-export default function CheckIcon(props: any) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      fill="none"
-      viewBox="0 0 16 16"
-      {...props}
-    >
-      <path
-        className="fill-current"
-        fillRule="evenodd"
-        d="M13.416 2.876a.75.75 0 0 1 .208 1.04l-6 9a.75.75 0 0 1-1.154.114l-4-4a.75.75 0 0 1 1.06-1.06l3.353 3.353 5.493-8.239a.75.75 0 0 1 1.04-.208Z"
-        clipRule="evenodd"
-      />
-    </svg>
-  );
-}
