--- src/assets/icons/CoinStackIcon.tsx
@@ -1,17 +1,0 @@
-export default function CoinStackIcon(props: any) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      fill="none"
-      viewBox="0 0 24 24"
-      {...props}
-    >
-      <path
-        className="fill-current"
-        d="M12 11c2.5 0 4.625-.392 6.375-1.175S21 8.1 21 7s-.875-2.042-2.625-2.825C16.625 3.392 14.5 3 12 3s-4.625.392-6.375 1.175S3 5.9 3 7s.875 2.042 2.625 2.825C7.375 10.608 9.5 11 12 11Zm0 2.5c.683 0 1.537-.07 2.563-.213a16.483 16.483 0 0 0 2.962-.687c.95-.317 1.767-.73 2.45-1.237C20.658 10.854 21 10.233 21 9.5V12c0 .733-.342 1.354-1.025 1.863-.683.508-1.5.92-2.45 1.237-.95.317-1.937.546-2.962.688-1.026.141-1.88.212-2.563.212-.683 0-1.537-.07-2.563-.213a16.483 16.483 0 0 1-2.962-.687c-.95-.317-1.767-.73-2.45-1.237C3.342 13.354 3 12.733 3 12V9.5c0 .733.342 1.354 1.025 1.863.683.508 1.5.92 2.45 1.237.95.317 1.938.546 2.963.688 1.024.141 1.879.212 2.562.212Zm0 5c.683 0 1.537-.07 2.563-.212a16.483 16.483 0 0 0 2.962-.688c.95-.317 1.767-.73 2.45-1.238C20.658 15.855 21 15.234 21 14.5V17c0 .733-.342 1.354-1.025 1.863-.683.508-1.5.92-2.45 1.237-.95.317-1.937.546-2.962.688-1.026.141-1.88.212-2.563.212-.683 0-1.537-.07-2.563-.212a16.483 16.483 0 0 1-2.962-.688c-.95-.317-1.767-.73-2.45-1.238C3.342 18.355 3 17.733 3 17v-2.5c0 .733.342 1.354 1.025 1.863.683.508 1.5.92 2.45 1.237.95.317 1.938.546 2.963.688 1.024.141 1.879.212 2.562.212Z"
-      />
-    </svg>
-  );
-}
