--- src/assets/icons/DecreaseIcon.tsx
@@ -1,19 +1,0 @@
-export default function DecreaseIcon(props: any) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="20"
-      height="20"
-      fill="none"
-      viewBox="0 0 20 20"
-      {...props}
-    >
-      <path
-        className="fill-current"
-        fillRule="evenodd"
-        d="M1.345 4.47a.75.75 0 0 1 1.06 0L7.5 9.564l3.041-3.041a.75.75 0 0 1 .985-.066 10.707 10.707 0 0 1 3.859 5.761l.178.665 1.746-3.024a.75.75 0 0 1 1.299.75l-2.652 4.593a.75.75 0 0 1-1.024.274l-4.593-2.652a.75.75 0 0 1 .75-1.299l3.025 1.747-.179-.667a9.207 9.207 0 0 0-2.813-4.541L8.03 11.155a.75.75 0 0 1-1.06 0L1.345 5.53a.75.75 0 0 1 0-1.06Z"
-        clipRule="evenodd"
-      />
-    </svg>
-  );
-}
