+++ src/assets/icons/Icon.tsx
@@ -1,0 +1,212 @@
+export function ArrowTriangleIcon(props: any) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="12"
+      height="13"
+      fill="none"
+      viewBox="0 0 12 13"
+      {...props}
+    >
+      <path
+        className="fill-current"
+        fillRule="evenodd"
+        d="M2.654 4.731A.375.375 0 0 1 3 4.5h6a.375.375 0 0 1 .265.64l-3 3a.375.375 0 0 1-.53 0l-3-3a.375.375 0 0 1-.081-.409Z"
+        clipRule="evenodd"
+      />
+    </svg>
+  );
+}
+
+export function CheckIcon(props: any) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="16"
+      height="16"
+      fill="none"
+      viewBox="0 0 16 16"
+      {...props}
+    >
+      <path
+        className="fill-current"
+        fillRule="evenodd"
+        d="M13.416 2.876a.75.75 0 0 1 .208 1.04l-6 9a.75.75 0 0 1-1.154.114l-4-4a.75.75 0 0 1 1.06-1.06l3.353 3.353 5.493-8.239a.75.75 0 0 1 1.04-.208Z"
+        clipRule="evenodd"
+      />
+    </svg>
+  );
+}
+
+export function CopyIcon(props: any) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="16"
+      height="16"
+      fill="none"
+      viewBox="0 0 16 16"
+      {...props}
+    >
+      <path
+        className="fill-current"
+        d="M9.333 14.667H2.667a1.29 1.29 0 0 1-1.334-1.334V6.667a1.29 1.29 0 0 1 1.334-1.334h2.666V2.667a1.29 1.29 0 0 1 1.334-1.334h6.666a1.29 1.29 0 0 1 1.334 1.334v6.666a1.29 1.29 0 0 1-1.334 1.334h-2.666v2.666a1.29 1.29 0 0 1-1.334 1.334Zm-6.666-8v6.666h6.666v-2.666H6.667a1.29 1.29 0 0 1-1.334-1.334V6.667H2.667Zm4-4v6.666h6.666V2.667H6.667Z"
+      />
+    </svg>
+  );
+}
+
+export function OutlinkIcon(props: any) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="16"
+      height="16"
+      fill="none"
+      viewBox="0 0 16 16"
+      {...props}
+    >
+      <path
+        className="fill-current"
+        fillRule="evenodd"
+        d="M9.9 2a.6.6 0 0 1 .6-.6H14a.6.6 0 0 1 .6.6v3.5a.6.6 0 1 1-1.2 0V3.449l-7.976 7.975a.6.6 0 0 1-.848-.848L12.55 2.6H10.5a.6.6 0 0 1-.6-.6ZM2.015 4.015A2.1 2.1 0 0 1 3.5 3.4H9a.6.6 0 0 1 0 1.2H3.5a.9.9 0 0 0-.9.9v7a.9.9 0 0 0 .9.9h7a.9.9 0 0 0 .9-.9V7a.6.6 0 1 1 1.2 0v5.5a2.1 2.1 0 0 1-2.1 2.1h-7a2.1 2.1 0 0 1-2.1-2.1v-7a2.1 2.1 0 0 1 .615-1.485Z"
+        clipRule="evenodd"
+      />
+    </svg>
+  );
+}
+
+export function DecreaseIcon(props: any) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="20"
+      height="20"
+      fill="none"
+      viewBox="0 0 20 20"
+      {...props}
+    >
+      <path
+        className="fill-current"
+        fillRule="evenodd"
+        d="M1.345 4.47a.75.75 0 0 1 1.06 0L7.5 9.564l3.041-3.041a.75.75 0 0 1 .985-.066 10.707 10.707 0 0 1 3.859 5.761l.178.665 1.746-3.024a.75.75 0 0 1 1.299.75l-2.652 4.593a.75.75 0 0 1-1.024.274l-4.593-2.652a.75.75 0 0 1 .75-1.299l3.025 1.747-.179-.667a9.207 9.207 0 0 0-2.813-4.541L8.03 11.155a.75.75 0 0 1-1.06 0L1.345 5.53a.75.75 0 0 1 0-1.06Z"
+        clipRule="evenodd"
+      />
+    </svg>
+  );
+}
+
+export function IncreaseIcon(props: any) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="20"
+      height="20"
+      fill="none"
+      viewBox="0 0 20 20"
+      {...props}
+    >
+      <path
+        className="fill-current"
+        fillRule="evenodd"
+        d="M12.566 5.18a.75.75 0 0 1 .97-.432l4.95 1.9a.75.75 0 0 1 .43.97l-1.9 4.95a.75.75 0 0 1-1.4-.537l1.252-3.261-.63.28a9.209 9.209 0 0 0-4.48 4.253.75.75 0 0 1-1.2.191L7.5 10.437 2.405 15.53a.75.75 0 0 1-1.06-1.06L6.97 8.845a.75.75 0 0 1 1.06 0l2.894 2.894a10.708 10.708 0 0 1 4.705-4.06l.629-.28-3.26-1.25a.75.75 0 0 1-.431-.97Z"
+        clipRule="evenodd"
+      />
+    </svg>
+  );
+}
+
+export function PlusIcon(props: any) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="16"
+      height="16"
+      fill="none"
+      viewBox="0 0 16 16"
+      {...props}
+    >
+      <path
+        className="fill-current"
+        d="M8.667 8.667v4H7.334v-4h-4V7.333h4v-4h1.333v4h4v1.334h-4Z"
+      />
+    </svg>
+  );
+}
+
+export function MinusIcon(props: any) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="16"
+      height="16"
+      fill="none"
+      viewBox="0 0 16 16"
+      {...props}
+    >
+      <path className="fill-current" d="M3.333 8.667V7.333h9.333v1.334H3.333Z" />
+    </svg>
+  );
+}
+
+export function BoosterIcon(props: any) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      fill="none"
+      viewBox="0 0 24 24"
+      {...props}
+    >
+      <path
+        className="fill-current"
+        fillRule="evenodd"
+        d="M12 20.667a8.667 8.667 0 1 0 0-17.334 8.667 8.667 0 0 0 0 17.334Zm0 2c5.89 0 10.666-4.776 10.666-10.667S17.891 1.333 12 1.333 1.333 6.11 1.333 12 6.109 22.667 12 22.667Z"
+        clipRule="evenodd"
+      />
+      <path
+        className="fill-current"
+        d="M9.867 17.333 10.4 13.6H7.733l4.8-6.933H13.6l-.533 4.266h3.2l-5.334 6.4H9.867Z"
+      />
+    </svg>
+  );
+}
+
+export function CoinStackIcon(props: any) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      fill="none"
+      viewBox="0 0 24 24"
+      {...props}
+    >
+      <path
+        className="fill-current"
+        d="M12 11c2.5 0 4.625-.392 6.375-1.175S21 8.1 21 7s-.875-2.042-2.625-2.825C16.625 3.392 14.5 3 12 3s-4.625.392-6.375 1.175S3 5.9 3 7s.875 2.042 2.625 2.825C7.375 10.608 9.5 11 12 11Zm0 2.5c.683 0 1.537-.07 2.563-.213a16.483 16.483 0 0 0 2.962-.687c.95-.317 1.767-.73 2.45-1.237C20.658 10.854 21 10.233 21 9.5V12c0 .733-.342 1.354-1.025 1.863-.683.508-1.5.92-2.45 1.237-.95.317-1.937.546-2.962.688-1.026.141-1.88.212-2.563.212-.683 0-1.537-.07-2.563-.213a16.483 16.483 0 0 1-2.962-.687c-.95-.317-1.767-.73-2.45-1.237C3.342 13.354 3 12.733 3 12V9.5c0 .733.342 1.354 1.025 1.863.683.508 1.5.92 2.45 1.237.95.317 1.938.546 2.963.688 1.024.141 1.879.212 2.562.212Zm0 5c.683 0 1.537-.07 2.563-.212a16.483 16.483 0 0 0 2.962-.688c.95-.317 1.767-.73 2.45-1.238C20.658 15.855 21 15.234 21 14.5V17c0 .733-.342 1.354-1.025 1.863-.683.508-1.5.92-2.45 1.237-.95.317-1.937.546-2.962.688-1.026.141-1.88.212-2.563.212-.683 0-1.537-.07-2.563-.212a16.483 16.483 0 0 1-2.962-.688c-.95-.317-1.767-.73-2.45-1.238C3.342 18.355 3 17.733 3 17v-2.5c0 .733.342 1.354 1.025 1.863.683.508 1.5.92 2.45 1.237.95.317 1.938.546 2.963.688 1.024.141 1.879.212 2.562.212Z"
+      />
+    </svg>
+  );
+}
+
+export function WaterdropIcon(props: any) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="16"
+      height="16"
+      fill="none"
+      viewBox="0 0 16 16"
+      {...props}
+    >
+      <path
+        className="fill-current"
+        fill-rule="evenodd"
+        d="m8.167 2-.277-.416.277.416Zm-.277-.416a.5.5 0 0 1 .56.004L8.168 2c.284-.412.285-.411.285-.411l.004.003.012.008a3.677 3.677 0 0 1 .178.133c.118.09.285.225.485.4.4.352.933.872 1.477 1.551C11.694 5.041 12.836 7.053 13 9.635v.032c0 1.365-.154 2.757-1.165 3.854-.965 1.048-2.284 1.646-3.668 1.646-1.384 0-2.702-.598-3.667-1.646-1.011-1.097-1.166-2.49-1.166-3.854l.001-.033c.175-2.632 1.316-4.643 2.401-5.99A11.903 11.903 0 0 1 7.21 2.117a9.287 9.287 0 0 1 .663-.52l.011-.008.004-.003h.002Z"
+        clip-rule="evenodd"
+      />
+    </svg>
+  );
+}
