--- src/assets/icons/IncreaseIcon.tsx
@@ -1,19 +1,0 @@
-export default function IncreaseIcon(props: any) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="20"
-      height="20"
-      fill="none"
-      viewBox="0 0 20 20"
-      {...props}
-    >
-      <path
-        className="fill-current"
-        fillRule="evenodd"
-        d="M12.566 5.18a.75.75 0 0 1 .97-.432l4.95 1.9a.75.75 0 0 1 .43.97l-1.9 4.95a.75.75 0 0 1-1.4-.537l1.252-3.261-.63.28a9.209 9.209 0 0 0-4.48 4.253.75.75 0 0 1-1.2.191L7.5 10.437 2.405 15.53a.75.75 0 0 1-1.06-1.06L6.97 8.845a.75.75 0 0 1 1.06 0l2.894 2.894a10.708 10.708 0 0 1 4.705-4.06l.629-.28-3.26-1.25a.75.75 0 0 1-.431-.97Z"
-        clipRule="evenodd"
-      />
-    </svg>
-  );
-}
