--- src/assets/icons/MinusIcon.tsx
@@ -1,14 +1,0 @@
-export default function MinusIcon(props: any) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      fill="none"
-      viewBox="0 0 16 16"
-      {...props}
-    >
-      <path className="fill-current" d="M3.333 8.667V7.333h9.333v1.334H3.333Z" />
-    </svg>
-  );
-}
