--- src/assets/icons/OutlinkIcon.tsx
@@ -1,19 +1,0 @@
-export default function OutlinkIcon(props: any) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      fill="none"
-      viewBox="0 0 16 16"
-      {...props}
-    >
-      <path
-        className="fill-current"
-        fillRule="evenodd"
-        d="M9.9 2a.6.6 0 0 1 .6-.6H14a.6.6 0 0 1 .6.6v3.5a.6.6 0 1 1-1.2 0V3.449l-7.976 7.975a.6.6 0 0 1-.848-.848L12.55 2.6H10.5a.6.6 0 0 1-.6-.6ZM2.015 4.015A2.1 2.1 0 0 1 3.5 3.4H9a.6.6 0 0 1 0 1.2H3.5a.9.9 0 0 0-.9.9v7a.9.9 0 0 0 .9.9h7a.9.9 0 0 0 .9-.9V7a.6.6 0 1 1 1.2 0v5.5a2.1 2.1 0 0 1-2.1 2.1h-7a2.1 2.1 0 0 1-2.1-2.1v-7a2.1 2.1 0 0 1 .615-1.485Z"
-        clipRule="evenodd"
-      />
-    </svg>
-  );
-}
