--- src/assets/icons/PlusIcon.tsx
@@ -1,17 +1,0 @@
-export default function PlusIcon(props: any) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      fill="none"
-      viewBox="0 0 16 16"
-      {...props}
-    >
-      <path
-        className="fill-current"
-        d="M8.667 8.667v4H7.334v-4h-4V7.333h4v-4h1.333v4h4v1.334h-4Z"
-      />
-    </svg>
-  );
-}
