--- src/assets/icons/WaterdropIcon.tsx
@@ -1,19 +1,0 @@
-export default function WaterdropIcon(props: any) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      fill="none"
-      viewBox="0 0 16 16"
-      {...props}
-    >
-      <path
-        className="fill-current"
-        fill-rule="evenodd"
-        d="m8.167 2-.277-.416.277.416Zm-.277-.416a.5.5 0 0 1 .56.004L8.168 2c.284-.412.285-.411.285-.411l.004.003.012.008a3.677 3.677 0 0 1 .178.133c.118.09.285.225.485.4.4.352.933.872 1.477 1.551C11.694 5.041 12.836 7.053 13 9.635v.032c0 1.365-.154 2.757-1.165 3.854-.965 1.048-2.284 1.646-3.668 1.646-1.384 0-2.702-.598-3.667-1.646-1.011-1.097-1.166-2.49-1.166-3.854l.001-.033c.175-2.632 1.316-4.643 2.401-5.99A11.903 11.903 0 0 1 7.21 2.117a9.287 9.287 0 0 1 .663-.52l.011-.008.004-.003h.002Z"
-        clip-rule="evenodd"
-      />
-    </svg>
-  );
-}
