+++ src/stories/molecule/AccountPopoverV3/index.tsx
@@ -2,7 +2,7 @@ import './style.css';
 
 import { Popover } from '@headlessui/react';
 import { PlusCircleIcon } from '@heroicons/react/24/outline';
-import ArrowTriangleIcon from '~/assets/icons/ArrowTriangleIcon';
+import { ArrowTriangleIcon } from '~/assets/icons/Icon';
 import { Avatar } from '~/stories/atom/Avatar';
 import { SkeletonElement } from '~/stories/atom/SkeletonElement';
 import { AccountPanelV3 } from '../AccountPanelV3';
