+++ src/stories/template/PoolMenu/index.tsx
@@ -1,4 +1,4 @@
-import ArrowTriangleIcon from '~/assets/icons/ArrowTriangleIcon';
+import { ArrowTriangleIcon } from '~/assets/icons/Icon';
 import '~/stories/atom/Tabs/style.css';
 import { Tag } from '~/stories/atom/Tag';
 import { usePoolMenu } from './hooks';
